backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
onnxruntime==1.17.0
opencv-python==4.9.0.80
numpy==1.26.3
supabase==2.3.4
httpx==0.26.0
pydantic==2.5.3
python-multipart==0.0.6
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.3
python-dotenv==1.0.0
